+++ website/urls.py
@@ -1,10 +1,10 @@
 from django.conf.urls import url
 
-from . import views
+# from . import views
 
 app_name = "website"
 urlpatterns = [
-    url(r'^$', views.index, name='index'),
+    url(r'^$', views.index.index, name='index'),
     url(r'^register$', views.register, name='register'),
     url(r'^login$', views.login_user, name='login'),
     url(r'^logout$', views.user_logout, name='logout'),
