+++ website/views.py
@@ -11,20 +11,7 @@ from django.template import RequestContext
 from website.forms import UserForm, ProductForm, AddPaymentForm
 from website.models import Product, Category, PaymentType, Order, UserOrder
 
-#display the index template
-def index(request):
-    '''
-    purpose: Shows last 20 products that have been added to the database
 
-    author: miriam rozenbuam
-
-    args: request
-
-    returns: (render): a view of the request, template to use, and product obj
-    ''' 
-    template_name = 'index.html'
-    newest_20_products =  Product.objects.all().order_by("-id")[:20]
-    return render(request, template_name, {'newest_20_products':newest_20_products})
 
 def register(request):
     """
